requirements.txt
reportlab
python-vlc
opencv-python
PyQt6
simpleaudio
watchdog
pymupdf
PyPDF2
pillow